+++ plugins/inputs/all/all.go
@@ -29,7 +29,7 @@ import (
 	_ "github.com/influxdata/telegraf/plugins/inputs/ipmi_sensor"
 	_ "github.com/influxdata/telegraf/plugins/inputs/jolokia"
 	_ "github.com/influxdata/telegraf/plugins/inputs/kafka_consumer"
-	_ "github.com/influxdata/telegraf/plugins/inputs/kafka_consumer_tls"
+	_ "github.com/influxdata/telegraf/plugins/inputs/kafka_consumer_plus"
 	_ "github.com/influxdata/telegraf/plugins/inputs/leofs"
 	_ "github.com/influxdata/telegraf/plugins/inputs/logparser"
 	_ "github.com/influxdata/telegraf/plugins/inputs/lustre2"
